requirements_github.txt
discord.py==2.3.2
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0